+++ src/contexts/GustoApiProvider/index.tsx
@@ -1,4 +1,3 @@
 export * from './GustoProvider'
-export * from './useGustoProvider'
 export { SDKI18next } from './SDKI18next'
 export * from './GustoProviderCustomUIAdapter'
